requirements.txt
numpy
pillow
imagehash
opencv-python
scikit-image
torch
lpips
open_clip_torch
